+++ src/internal/DelayDSPKernel.h
@@ -22,7 +22,7 @@ public:
     DelayDSPKernel(double maxDelayTime, float sampleRate);
     virtual ~DelayDSPKernel() {}
 
-    virtual void process(ContextRenderLock &, const float * source, float * destination, size_t framesToProcess) override;
+    virtual void process(ContextRenderLock &, const float * source, float * destination, int framesToProcess) override;
     virtual void reset() override;
 
     double maxDelayTime() const { return m_maxDelayTime; }
