+++ src/internal/EqualPowerPanner.h
@@ -19,7 +19,7 @@ public:
 
     virtual void pan(ContextRenderLock &, double azimuth, double elevation,
                      const AudioBus * inputBus, AudioBus * outputBuf,
-                     size_t framesToProcess) override;
+                     int framesToProcess) override;
 
     virtual void reset() override { m_isFirstRender = true; }
